requirements.txt
    Flask
    pymongo
    opencv-python
    facenet-pytorch
    joblib
    pillow
    scikit-learn
    torchvision